use io performance test utilities

+++ sdks/java/io/google-cloud-platform/src/main/java/org/apache/beam/sdk/io/gcp/bigquery/BigQueryOptions.java
@@ -168,10 +168,10 @@ public interface BigQueryOptions
   @Internal
   @Description(
       "Intended for testing purposes only. If set, the Storage API sink will periodically crash at this interval.")
-  @Default.Long(-1)
-  Long getCrashStorageApiWriteEverySeconds();
+  @Default.Integer(-1)
+  Integer getCrashStorageApiWriteEverySeconds();
 
-  void setCrashStorageApiWriteEverySeconds(Long value);
+  void setCrashStorageApiWriteEverySeconds(Integer value);
 
   @Description(
       "If set, BigQueryIO.Read will use the StreamBundle based"
